Reorganize handler methods
